+++ R/plotting.R
@@ -11,7 +11,7 @@
 #' @param composite_rug A boolean option to plot a rug on the bottom of the plot. Default is FALSE. Note that composite_rug and facet_group, facet_id cannot be used in the same plot. You must choose facets or a composite rug.
 #' @param filter_prop An optional argument if the user chooses to include a composite rug in their plot. This is passed to \code{composite}. See this function for details.
 #' @param filter_min_rec An optional argument if the user chooses to include a composite rug in their plot. This is passed to \code{composite}. See this function for details.
-#' @param filter_min_scarred An optional argument if the user chooses to include a composite rug in their plot. This is passed to \code{composite}. See this function for details.
+#' @param filter_min_events An optional argument if the user chooses to include a composite rug in their plot. This is passed to \code{composite}. See this function for details.
 #' @param injury_event Boolean indicating whether injuries should be considered recorders. This is passed to \code{composite}. See this function for details.
 #' @param plot_legend A boolean option allowing the user to choose whether a legend is included in the plot or not. Default is FALSE.
 #' @param event_size An optional numeric vector that adjusts the size of fire event symbols on the plot. Default is \code{c("Scar" = 4, "Injury" = 2, "Pith/Bark" = 1.5)}.
@@ -52,7 +52,7 @@
 plot_demograph <- function(x, color_group, color_id, facet_group, facet_id,
                        facet_type="grid", ylabels=TRUE, yearlims=FALSE,
                        composite_rug=FALSE, filter_prop=0.25, filter_min_rec=2,
-                       filter_min_scarred=1, injury_event= FALSE, plot_legend=FALSE,
+                       filter_min_events=1, injury_event= FALSE, plot_legend=FALSE,
                        event_size=c("Scar" = 4, "Injury" = 2, "Pith/Bark" = 1.5),
                        rugbuffer_size=2, rugdivide_pos=2) {
 # TODO: Merge ends and events into a single df. with a factor to handle the
@@ -129,7 +129,7 @@ plot_demograph <- function(x, color_group, color_id, facet_group, facet_id,
   if (composite_rug) {
     comp <- composite(x, filter_prop = filter_prop,
                         filter_min_rec = filter_min_rec,
-                        filter_min_scarred = filter_min_scarred,
+                        filter_min_events = filter_min_events,
                         injury_event = injury_event)
     p <- (p + ggplot2::geom_rug(data = rings[rings$year %in% get_event_years(comp, injury_event = injury_event)[['COMP']], ],
                        sides = "b", color = "black")